% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{best_subset_fitting}
\alias{best_subset_fitting}
\title{Best Subset Fitting}
\usage{
best_subset_fitting(mut_data, prior_knowledge_subset)
}
\arguments{
\item{mut_data}{The mutation data to be analyzed.}

\item{prior_knowledge_subset}{A subset of prior knowledge to be used in the fitting.}
}
\value{
A tibble containing the contribution data of the best subset fitting.
}
\description{
This function fits mutation data to the best subset based on prior knowledge.
}
\details{
This function fits mutation data to the best subset based on the provided prior knowledge subset.
It uses the method of fitting to signatures strictly with a maximum delta of 0.0001 and the backward method.
The function then generates a dummy plot to pull out the total contribution data, which is used to reorder bars
in the plot. The contribution data is returned as a tibble.
}
\examples{
\dontrun{
# Example usage:
# Fit mutation data to the best subset based on prior knowledge:
contribution_data = best_subset_fitting(mut_data, prior_knowledge_subset)
}
}
