% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{best_subset_plot}
\alias{best_subset_plot}
\title{Master Function to Calculate, Sort, and Replot Results}
\usage{
best_subset_plot(mut_data, sample_names, prior_knowledge_subset)
}
\arguments{
\item{mut_data}{The mutation data used for analysis.}

\item{sample_names}{A character vector specifying the sample names.}

\item{prior_knowledge_subset}{A subset of prior knowledge to be used in the analysis.}
}
\description{
This function serves as a master routine to calculate, sort, and replot results based on mutation data.
}
\details{
This function performs the following steps:
\enumerate{
\item Calculates the best subset fitting based on mutation data and prior knowledge subset.
\item Reorders the resulting dataframe based on the specified sample names.
\item Plots the reordered contribution of each subset.
}
}
\examples{
\dontrun{
# Example usage:
# Calculate, sort, and replot results based on mutation data:
best_subset_plot(mut_data, sample_names, prior_knowledge_subset)
}
}
