% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapped_background_correction.R
\name{bootstrap_routine}
\alias{bootstrap_routine}
\title{Master Bootstrap Routine}
\usage{
bootstrap_routine(input_matrix)
}
\arguments{
\item{input_matrix}{The input matrix to be bootstrapped.}
}
\value{
A hash containing bootstrapped matrices obtained from the input matrix.
}
\description{
This function serves as a master routine for bootstrapping a given input matrix.
}
\details{
This function takes an input matrix and performs the following steps:
\enumerate{
\item Extracts the input matrix into a hash.
\item Prepares bootstrap iterations in the hash.
\item Performs bootstrapping using the prepared iterations.
The function then returns a hash containing bootstrapped matrices derived from the input matrix.
}
}
\examples{
\dontrun{
# Example usage:
# Perform bootstrapping routine on an input matrix:
bootstrapped_results = bootstrap_routine(input_matrix)
}
}
