% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{check_all_zeros}
\alias{check_all_zeros}
\title{Check if Column Contains All Zeros}
\usage{
check_all_zeros(matrix, column_index)
}
\arguments{
\item{matrix}{The matrix in which to check the column.}

\item{column_index}{The index of the column to be checked for all zeros.}
}
\value{
TRUE if the specified column contains all zeros, FALSE otherwise.
}
\description{
This function checks if a specified column in a matrix contains all zeros.
}
\details{
This function takes a matrix and a column index as input. It checks if the column specified
by the index contains all zeros. If all elements in the column are zero, the function returns TRUE; otherwise,
it returns FALSE.
}
\examples{
\dontrun{
# Example usage:
# Check if a column contains all zeros:
is_all_zeros = check_all_zeros(matrix_data, 2)
}
}
