% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{check_sum_less_than_TH}
\alias{check_sum_less_than_TH}
\title{Check if Sum of Column is Less Than Threshold}
\usage{
check_sum_less_than_TH(matrix, column_index, threshold)
}
\arguments{
\item{matrix}{The matrix in which to check the column sum.}

\item{column_index}{The index of the column to be checked for its sum.}

\item{threshold}{The threshold value to compare against the sum of the column.}
}
\value{
TRUE if the sum of the specified column is less than the threshold, FALSE otherwise.
}
\description{
This function checks if the sum of a specified column in a matrix is less than a given threshold.
}
\details{
This function takes a matrix, a column index, and a threshold value as input. It calculates
the sum of the column specified by the index and checks if it is less than the threshold. If the sum
is less than the threshold, the function returns TRUE; otherwise, it returns FALSE.
}
\examples{
\dontrun{
# Example usage:
# Check if the sum of a column is less than a threshold:
is_sum_less_than_threshold = check_sum_less_than_TH(matrix_data, 2, 1000)
}
}
