% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapped_background_correction.R
\name{gather_replicates_into_hash}
\alias{gather_replicates_into_hash}
\title{Gather Replicate Matrices into a Hash for Bootstrapping}
\usage{
gather_replicates_into_hash(mutation_hash, num_replica)
}
\arguments{
\item{mutation_hash}{A hash object where each key contains a matrix to be replicated.}

\item{num_replica}{An integer specifying the number of replications for each matrix in the hash.}
}
\value{
A hash object where each key contains a replicated matrix.
}
\description{
This function creates a hash of replicate matrices, ready for bootstrapping.
The number of bootstrapped samples is equal to the specified number of replications.
}
\details{
This function iterates over each matrix in the input hash, replicates its columns
the specified number of times, and stores the resulting matrix in a new hash. This is useful
for preparing data for bootstrapping.
}
\examples{
\dontrun{
library(hash)
mat1 = matrix(1:6, nrow = 2, dimnames = list(NULL, c("A", "B", "C")))
mat2 = matrix(7:12, nrow = 2, dimnames = list(NULL, c("D", "E", "F")))
mutation_hash = hash()
mutation_hash[["clone1"]] = mat1
mutation_hash[["clone2"]] = mat2
replicated_hash = gather_replicates_into_hash(mutation_hash, 3)
print(replicated_hash)
}

}
