% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapped_background_correction.R
\name{hash_to_average_matrix}
\alias{hash_to_average_matrix}
\title{Convert Hash of Averaged Results to Matrix}
\usage{
hash_to_average_matrix(averaged_samples_hash)
}
\arguments{
\item{averaged_samples_hash}{A hash object where each key contains a single column of averaged results (row-wise average).}
}
\value{
A matrix with averaged results, where each column corresponds to a key in the input hash.
}
\description{
This function converts a hash of averaged results into a matrix.
}
\details{
This function iterates over each item in the input hash, constructs a data frame from the averaged results,
and then converts this data frame into a matrix. The row names of the resulting matrix are set to match
the row names of the original mutation matrix (\code{mut_mat}).
}
\examples{
\dontrun{
library(hash)
library(S4Vectors)
averaged_results = hash()
averaged_results[["sample1"]] = runif(5)
averaged_results[["sample2"]] = runif(5)
mut_mat = matrix(1:10, nrow = 5)  # Example mutation matrix to provide row names
average_matrix = hash_to_average_matrix(averaged_results)
print(average_matrix)
}
}
