% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapped_background_correction.R
\name{hash_to_matrix}
\alias{hash_to_matrix}
\title{Convert Hash of Bootstrapped Results to Matrix}
\usage{
hash_to_matrix(input_hash)
}
\arguments{
\item{input_hash}{A hash object where each key contains a list of bootstrapped matrices.}
}
\value{
A matrix containing all bootstrapped samples concatenated column-wise.
}
\description{
This function converts a hash of bootstrapped results into a matrix for error bars.
}
\details{
This function iterates over each list of bootstrapped matrices in the input hash,
combines them into a single data frame, and converts this data frame into a matrix.
}
\examples{
\dontrun{
library(hash)
bootstrapped_results = hash()
bootstrapped_results[["sample1"]] = matrix(runif(20), nrow = 5)
bootstrapped_results[["sample2"]] = matrix(runif(20), nrow = 5)
bootstrap_matrix = hash_to_matrix(bootstrapped_results)
print(bootstrap_matrix)
}
}
