% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapped_background_correction.R
\name{matrix_to_hash}
\alias{matrix_to_hash}
\title{Convert Matrix Columns to Hash}
\usage{
matrix_to_hash(input_matrix)
}
\arguments{
\item{input_matrix}{A matrix with named columns.}
}
\value{
A hash object where each key is a column name and each value is a vector
corresponding to the matrix column.
}
\description{
This function takes a matrix as input and stores each column of the matrix
in a hash, using the column names as keys.
}
\details{
This function is useful for separating out all the clones in a matrix
and storing each with its own key.
}
\examples{
\dontrun{
library(hash)
mat = matrix(1:9, nrow = 3, dimnames = list(NULL, c("A", "B", "C")))
result_hash = matrix_to_hash(mat)
print(result_hash)
}

}
