% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapped_background_correction.R
\name{median_bootstrapped_samples}
\alias{median_bootstrapped_samples}
\title{Compute Median of Bootstrapped Samples}
\usage{
median_bootstrapped_samples(hash_of_bootstrap_results)
}
\arguments{
\item{hash_of_bootstrap_results}{A hash object where each key contains a list of bootstrapped matrices.}
}
\value{
A hash object where each key contains the median values of the bootstrapped matrices.
}
\description{
This function pulls out the median (a measure of central tendency) of bootstrapped results
stored in a hash.
}
\details{
This function iterates over each list of bootstrapped matrices in the input hash,
computes the median for each row across the matrices, and stores the resulting medians in a new hash.
}
\examples{
\dontrun{
library(hash)
bootstrapped_results = hash()
bootstrapped_results[["sample1"]] = matrix(runif(20), nrow=5)
bootstrapped_results[["sample2"]] = matrix(runif(20), nrow=5)
median_results = median_bootstrapped_samples(bootstrapped_results)
print(median_results)
}
}
