% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{order_sample_names}
\alias{order_sample_names}
\title{Reorder Sample Names Based on Sorting Index}
\usage{
order_sample_names(sorting_index, sample_names)
}
\arguments{
\item{sorting_index}{A data frame containing a column 'sample_ID' specifying the desired order of sample names.}

\item{sample_names}{A character vector containing the original sample names to be reordered.}
}
\value{
A character vector containing the sample names reordered according to the sorting index.
}
\description{
This function reorders sample names based on a sorting index. The sorting index specifies the desired order
of sample names, typically based on genotype, day, and clone as serialized in the sample_mapping.csv file.
}
\details{
This function reorders the sample names according to the specified sorting index. It matches the
sample names in the sorting index with the original sample names and reorders them accordingly. Sample names
not found in the sorting index are omitted from the result.
}
\examples{
\dontrun{
# Example usage:
# Reorder sample names based on a sorting index:
sorted_names = order_sample_names(sorting_index_df, sample_names_vector)
}
}
