% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_apobec_heatmap}
\alias{plot_apobec_heatmap}
\title{Make Zoomed-in Heatmap of APOBEC-Related Signatures}
\usage{
plot_apobec_heatmap(refit_model, sample_selection, cell_line)
}
\arguments{
\item{refit_model}{The refitted model containing contribution data.}

\item{sample_selection}{A vector specifying the samples to be included in the heatmap.}

\item{cell_line}{A character string specifying the cell line for the plot title.}
}
\value{
A ggplot object representing the zoomed-in heatmap of APOBEC-related signatures.
}
\description{
This function creates a zoomed-in heatmap of APOBEC-related signatures.
}
\details{
This function takes a refitted model containing contribution data, a vector specifying
the samples to be included in the heatmap, and a character string specifying the cell line for
the plot title. It extracts contribution data for APOBEC-related signatures, creates a melted
dataframe, and generates a heatmap using ggplot2 library. The colors of the heatmap are based on
a gradient color palette. The function returns the ggplot object representing the zoomed-in heatmap
of APOBEC-related signatures.
}
\examples{
\dontrun{
# Example usage:
# Create a zoomed-in heatmap of APOBEC-related signatures:
plot_apobec_heatmap(refit_model, sample_selection, cell_line)
}
}
