% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_reordered_contribution}
\alias{plot_reordered_contribution}
\title{Plot Reordered Contribution}
\usage{
plot_reordered_contribution(reordered_contribution)
}
\arguments{
\item{reordered_contribution}{A dataframe containing reordered contribution data.}
}
\value{
A ggplot object representing the contribution by signatures.
}
\description{
This function plots the reordered contribution data.
}
\details{
This function takes a dataframe containing reordered contribution data as input. It generates
a bar plot using ggplot2 library, where the x-axis represents the samples, the y-axis represents the
contribution, and the bars are filled by signature types. The colors of signature types are manually specified.
The function returns the ggplot object representing the contribution by signatures.
}
\examples{
\dontrun{
# Example usage:
# Plot reordered contribution data:
plot_reordered_contribution(reordered_contribution)
}
}
