% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_specific_context}
\alias{plot_specific_context}
\title{Visualize Specific Sequence Context by Genotype}
\usage{
plot_specific_context(input_data, cell_line, context_name)
}
\arguments{
\item{input_data}{The input data to be reshaped and plotted.}

\item{cell_line}{A character string specifying the cell line for the plot title.}

\item{context_name}{A character string specifying the name of the specific sequence context.}
}
\value{
A ggplot object representing the visualization of the specific sequence context by genotype.
}
\description{
This function reshapes and plots data to visualize a specific sequence context by genotype.
}
\details{
This function takes input data, melts it to a long format using the 'melt' function from the
'reshape2' package, and then creates a bar plot using ggplot2 library. Each bar represents the percentage
of mutation contexts for a specific genotype. The x-axis represents the samples, the y-axis represents
the percentage, and bars are filled by variable. The function returns the ggplot object representing
the visualization of the specific sequence context by genotype.
}
\examples{
\dontrun{
# Example usage:
# Visualize SBS2 specific sequence context by genotype:
plot_specific_context(input_data, cell_line, context_name)
}
}
