% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapped_background_correction.R
\name{prepare_bootstrap_iterations}
\alias{prepare_bootstrap_iterations}
\title{Replicate Columns in a Matrix}
\usage{
prepare_bootstrap_iterations(input_matrix, num_replica)
}
\arguments{
\item{input_matrix}{A matrix whose columns will be replicated.}

\item{num_replica}{An integer specifying the number of times to replicate the columns of the input matrix.}
}
\value{
A matrix with replicated columns.
}
\description{
This helper function replicates the columns of an input matrix a specified number of times.
It is useful for preparing matrices for bootstrapping.
}
\details{
The function takes an input matrix and replicates its columns, creating a larger matrix
with repeated columns. This is useful for certain statistical techniques such as bootstrapping.
}
\examples{
\dontrun{
mat = matrix(1:6, nrow = 2, dimnames = list(NULL, c("A", "B", "C")))
boot_matrix = prepare_bootstrap_iterations(mat, 3)
print(boot_matrix)
}

}
