% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{reorder_dataframe}
\alias{reorder_dataframe}
\title{Reorder Dataframe Based on Sample Names}
\usage{
reorder_dataframe(df, sample_names)
}
\arguments{
\item{df}{The dataframe to be reordered.}

\item{sample_names}{A character vector specifying the desired order of sample names.}
}
\value{
A reordered dataframe based on sample names.
}
\description{
This function reorders a dataframe based on specified sample names.
}
\details{
This function takes a dataframe and a vector of sample names as input. It converts the 'Sample'
column of the dataframe to a factor with the specified order of sample names and removes duplicate rows.
The dataframe is then reordered based on the 'Signature' column and the new order of 'Sample'. The reordered
dataframe is returned as the result.
}
\examples{
\dontrun{
# Example usage:
# Reorder a dataframe based on sample names:
reordered_df = reorder_dataframe(df, sample_names)
}
}
