% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{reorder_matrix_columns}
\alias{reorder_matrix_columns}
\title{Reorder Matrix Columns According to Sample Name Vector}
\usage{
reorder_matrix_columns(matrix_data, sample_vector)
}
\arguments{
\item{matrix_data}{The matrix whose columns are to be reordered.}

\item{sample_vector}{A character vector specifying the desired order of sample names.}
}
\value{
A matrix with columns reordered according to the provided sample name vector.
}
\description{
This function reorders the columns of a matrix according to a given sample name vector.
}
\details{
This function takes a matrix and a sample name vector as input. It matches the sample names
in the sample vector with the column names of the matrix and reorders the columns accordingly. If a
sample name in the vector does not match any column name in the matrix, it is omitted from the result.
}
\examples{
\dontrun{
# Example usage:
# Reorder matrix columns according to a sample name vector:
reordered_matrix = reorder_matrix_columns(matrix_data, sample_vector)
}
}
