% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{subset_matrix_by_samples}
\alias{subset_matrix_by_samples}
\title{Subset Mutational Matrix by Sample Names}
\usage{
subset_matrix_by_samples(original_matrix, sample_names)
}
\arguments{
\item{original_matrix}{The original mutational matrix to be subsetted.}

\item{sample_names}{A character vector specifying the sample names to be included in the subset.}
}
\value{
A subset of the original mutational matrix containing only the specified sample names.
}
\description{
This function subsets a mutational matrix based on a vector of sample names.
}
\details{
This function takes an original mutational matrix and a vector of sample names as input.
It subsets the original matrix to include only the columns corresponding to the specified sample names.
If a sample name in the vector does not match any column name in the matrix, it is omitted from the result.
}
\examples{
\dontrun{
# Example usage:
# Subset a mutational matrix by a vector of sample names:
subsetted_matrix = subset_matrix_by_samples(original_matrix, sample_names)
}
}
